Move `metrics::Serve` into its own module
With this change, metrics/mod.rs now contains only metrics types.
